notification no data, notification non logged click, my trips non logged click

+++ TripsMan/Modules/Other/My Trips/Model/Data Model/HotelTripDetails.swift
@@ -18,7 +18,7 @@ struct HotelTripDetailsData: Codable {
 // MARK: - TripDetails
 struct HotelTripDetails: Codable {
     let bookingDate: String
-    let totalAmount: Int
+    let totalAmount: Double
     let bookingFrom, bookingTo: String
     let totalCharge, tax, serviceCharge, discount: Int
     let totalPrice, totalGuest: Int
